+++ app/controllers/people_controller.rb
@@ -3,8 +3,8 @@
 #   the COPYRIGHT file.
 
 class PeopleController < ApplicationController
-  before_action :authenticate_user!, except: [:show, :stream]
-  before_action :find_person, only: [:show, :stream, :hovercard]
+  before_action :authenticate_user!, except: %i(show stream hovercard)
+  before_action :find_person, only: %i(show stream hovercard)
 
   respond_to :html, :except => [:tag_index]
   respond_to :json, :only => [:index, :show]