+++ spec/federation_callbacks_spec.rb
@@ -1,4 +1,5 @@
 require "spec_helper"
+require "diaspora_federation/test"
 
 describe "diaspora federation callbacks" do
   describe ":fetch_person_for_webfinger" do
@@ -12,6 +13,7 @@ describe "diaspora federation callbacks" do
       expect(wf.profile_url).to eq(person.profile_url)
       expect(wf.atom_url).to eq(person.atom_url)
       expect(wf.salmon_url).to eq(person.receive_url)
+      expect(wf.subscribe_url).to eq(AppConfig.url_to("/people?q={uri}"))
       expect(wf.guid).to eq(person.guid)
       expect(wf.public_key).to eq(person.serialized_public_key)
     end
@@ -145,6 +147,153 @@ describe "diaspora federation callbacks" do
         expect(profile_entity.first_name).to eq(profile.first_name)
         expect(profile_entity.last_name).to eq(profile.last_name)
       end
+    end
+  end
+
+  let(:local_person) { FactoryGirl.create(:user).person }
+  let(:remote_person) { FactoryGirl.create(:person) }
+  let(:post_by_a_local_person) { FactoryGirl.create(:status_message, author: local_person) }
+  let(:post_by_a_remote_person) { FactoryGirl.create(:status_message, author: remote_person) }
+
+  describe ":fetch_private_key_by_diaspora_id" do
+    it "returns a private key for a local user" do
+      key = DiasporaFederation.callbacks.trigger(:fetch_private_key_by_diaspora_id, local_person.diaspora_handle)
+      expect(key).to be_a(OpenSSL::PKey::RSA)
+      expect(key.to_s).to eq(local_person.owner.serialized_private_key)
+    end
+
+    it "returns nil for a remote user" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_private_key_by_diaspora_id, remote_person.diaspora_handle)
+      ).to be_nil
+    end
+
+    it "returns nil for an unknown id" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_private_key_by_diaspora_id, FactoryGirl.generate(:diaspora_id))
+      ).to be_nil
+    end
+  end
+
+  describe ":fetch_author_private_key_by_entity_guid" do
+    it "returns a private key for a post by a local user" do
+      key = DiasporaFederation.callbacks.trigger(:fetch_author_private_key_by_entity_guid,
+                                                 "Post", post_by_a_local_person.guid)
+      expect(key).to be_a(OpenSSL::PKey::RSA)
+      expect(key.to_s).to eq(post_by_a_local_person.author.owner.serialized_private_key)
+    end
+
+    it "returns nil for a post by a remote user" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_author_private_key_by_entity_guid,
+                                             "Post", post_by_a_remote_person.guid)
+      ).to be_nil
+    end
+
+    it "returns nil for an unknown post" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_author_private_key_by_entity_guid,
+                                             "Post", FactoryGirl.generate(:guid))
+      ).to be_nil
+    end
+  end
+
+  describe ":fetch_public_key_by_diaspora_id" do
+    it "returns a public key for a person" do
+      key = DiasporaFederation.callbacks.trigger(:fetch_public_key_by_diaspora_id, remote_person.diaspora_handle)
+      expect(key).to be_a(OpenSSL::PKey::RSA)
+      expect(key.to_s).to eq(remote_person.serialized_public_key)
+    end
+
+    it "returns nil for an unknown person" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_public_key_by_diaspora_id, FactoryGirl.generate(:diaspora_id))
+      ).to be_nil
+    end
+  end
+
+  describe ":fetch_author_public_key_by_entity_guid" do
+    it "returns a public key for a known post" do
+      key = DiasporaFederation.callbacks.trigger(:fetch_author_public_key_by_entity_guid,
+                                                 "Post", post_by_a_remote_person.guid)
+      expect(key).to be_a(OpenSSL::PKey::RSA)
+      expect(key.to_s).to eq(post_by_a_remote_person.author.serialized_public_key)
+    end
+
+    it "returns nil for an unknown post" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_author_public_key_by_entity_guid,
+                                             "Post", FactoryGirl.generate(:guid))
+      ).to be_nil
+    end
+  end
+
+  describe ":entity_author_is_local?" do
+    it "returns true for a post by a local user" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:entity_author_is_local?, "Post", post_by_a_local_person.guid)
+      ).to be_truthy
+    end
+
+    it "returns false for a post by a remote user" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:entity_author_is_local?, "Post", post_by_a_remote_person.guid)
+      ).to be_falsey
+    end
+
+    it "returns false for a unknown post" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:entity_author_is_local?, "Post", FactoryGirl.generate(:diaspora_id))
+      ).to be_falsey
+    end
+  end
+
+  describe ":fetch_entity_author_id_by_guid" do
+    it "returns id for a existing guid" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_entity_author_id_by_guid, "Post", post_by_a_remote_person.guid)
+      ).not_to eq(post_by_a_remote_person.author_id)
+    end
+
+    it "returns nil for a non-existing guid" do
+      expect(
+        DiasporaFederation.callbacks.trigger(:fetch_entity_author_id_by_guid, "Post", FactoryGirl.generate(:guid))
+      ).to be_nil
+    end
+  end
+
+  describe ":queue_public_receive" do
+    it "enqueues a ReceiveUnencryptedSalmon job" do
+      xml = "<diaspora/>"
+      expect(Workers::ReceiveUnencryptedSalmon).to receive(:perform_async).with(xml)
+
+      DiasporaFederation.callbacks.trigger(:queue_public_receive, xml)
+    end
+  end
+
+  describe ":queue_private_receive" do
+    let(:xml) { "<diaspora/>" }
+
+    it "returns true if the user is found" do
+      result = DiasporaFederation.callbacks.trigger(:queue_private_receive, alice.person.guid, xml)
+      expect(result).to be_truthy
+    end
+
+    it "enqueues a ReceiveEncryptedSalmon job" do
+      expect(Workers::ReceiveEncryptedSalmon).to receive(:perform_async).with(alice.id, xml)
+
+      DiasporaFederation.callbacks.trigger(:queue_private_receive, alice.person.guid, xml)
+    end
+
+    it "returns false if the no user is found" do
+      person = FactoryGirl.create(:person)
+      result = DiasporaFederation.callbacks.trigger(:queue_private_receive, person.guid, xml)
+      expect(result).to be_falsey
+    end
+
+    it "returns false if the no person is found" do
+      result = DiasporaFederation.callbacks.trigger(:queue_private_receive, "2398rq3948yftn", xml)
+      expect(result).to be_falsey
     end
   end
 end